+++ src/decorators/orm/index.ts
@@ -3,6 +3,7 @@
 export * from './associations';
 
 export * from './counter.decorator';
+export * from './flag.decorator';
 export * from './model.decorator';
 export * from './pk.decorator';
 export * from './prop.decorator';